# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TrainSystem",
    "contact": {},
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
...
  "paths": {
    "/stations/new": {
      "post": {
        "tags": [
          "Station"
        ],
        "summary": "station",
        "description": "station",
        "operationId": "station",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/stations_new_body"
              },
              "examples": {
                "station": {
                  "value": {
                    "name": "Mansoura"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "stations_new_body": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mansoura"
          }
        }
      },
...
    }
  }
}